# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Football League API documentation for the Football League service.",
        "title": "Football League API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/fixtures/{fixtureId}": {
            "put": {
                "description": "Update the fixture's status and scores by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fixtures"
                ],
                "summary": "Update Fixture",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fixture ID",
                        "name": "fixtureId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fixture Update Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateFixtureDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fixture updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or missing fixture ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Fixture not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/fixtures/{gameweekId}": {
            "get": {
                "description": "Get Fixtures by a Gameweek Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fixtures"
                ],
                "summary": "Get Fixtures by Gameweek",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gameweek",
                        "name": "gameweekId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Fixture"
                            }
                        }
                    }
                }
            }
        },
        "/api/standings": {
            "get": {
                "description": "Retrieve team statistics for all teams",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Standings"
                ],
                "summary": "Get standings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TeamStatistics"
                            }
                        }
                    }
                }
            }
        },
        "/api/teams": {
            "get": {
                "description": "Retrieve details of all teams",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Get all teams",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Team"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.UpdateFixtureDTO": {
            "type": "object",
            "properties": {
                "awayScore": {
                    "type": "integer"
                },
                "homeScore": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.FixtureStatus"
                }
            }
        },
        "models.Fixture": {
            "type": "object",
            "properties": {
                "awayScore": {
                    "type": "integer"
                },
                "awayTeam": {
                    "description": "References Team.ID",
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "gameweekId": {
                    "type": "integer"
                },
                "homeScore": {
                    "type": "integer"
                },
                "homeTeam": {
                    "description": "References Team.ID",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.FixtureStatus"
                }
            }
        },
        "models.FixtureStatus": {
            "type": "string",
            "enum": [
                "Played",
                "Live",
                "Upcoming"
            ],
            "x-enum-varnames": [
                "StatusPlayed",
                "StatusLive",
                "StatusUpcoming"
            ]
        },
        "models.Team": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "stadium": {
                    "type": "string"
                },
                "stadiumCapacity": {
                    "type": "integer"
                }
            }
        },
        "models.TeamStatistics": {
            "type": "object",
            "properties": {
                "draws": {
                    "type": "integer"
                },
                "goalDifference": {
                    "type": "integer"
                },
                "goalsConceded": {
                    "type": "integer"
                },
                "goalsScored": {
                    "type": "integer"
                },
                "lastUpdated": {
                    "type": "string"
                },
                "losses": {
                    "type": "integer"
                },
                "numberOfGames": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "teamId": {
                    "description": "FK to Teams",
                    "type": "string"
                },
                "wins": {
                    "type": "integer"
                }
            }
        }
    }
}